requirements.txt
aiohttp==3.8.4
python-socketio==5.7.2
python-dotenv==1.0.0